cashcog-be/requirements.txt
click==7.0
pytest==5.3.5
sqlalchemy==1.3.13
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
alembic==1.4.1
psycopg2==2.8.4
pycodestyle==2.5.0
python-dotenv==0.10.3
Flask-GraphQL==2.0.1
graphene==2.1.8
loguru==0.4.1
factory_boy==2.12.0
validate_email==1.3
py3dns==3.2.0
pytest-mock==2.0.0
Flask-Cors==3.0.8
graphene-sqlalchemy==2.2.2
coverage==4.5.3